Move all windows into TabletDriverGUI.Windows namespace
