+++ TabletDriverGUI/MainWindow.Areas.cs
@@ -6,6 +6,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 using System.Windows.Threading;
+using TabletDriverGUI.Windows;
 
 namespace TabletDriverGUI
 {
