+++ TabletDriverGUI/MainWindow.Settings.cs
@@ -9,6 +9,7 @@ using System.Windows.Controls;
 using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Threading;
+using TabletDriverGUI.Windows;
 
 namespace TabletDriverGUI
 {
