+++ TabletDriverGUI/Windows/WindowTabletView.xaml.cs
@@ -8,7 +8,7 @@ using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 
-namespace TabletDriverGUI
+namespace TabletDriverGUI.Windows
 {
 
     /// <summary>
